finalise Campaign Manager view
