+++ intuita-webview/src/campaignManager/ListView/style.module.css
@@ -1,0 +1,3 @@
+.container {
+	width: 100%;
+}
