--- intuita-webview/src/campaignManager/TreeItem/style.module.css
@@ -1,102 +1,0 @@
-.root {
-	display: flex;
-	height: 22px;
-	line-height: 22px;
-	flex: 1;
-	text-overflow: ellipsis;
-	overflow: hidden;
-	flex-wrap: nowrap;
-	padding-left: 3px;
-	padding-right: 12px;
-}
-
-.root:hover {
-	background-color: var(--vscode-list-hoverBackground);
-}
-
-.root .actions {
-	display: none;
-}
-
-.root:hover .actions {
-	display: flex;
-}
-
-.focused,
-.focused:hover {
-	background-color: var(--vscode-list-activeSelectionBackground);
-}
-
-.icon {
-	min-width: 22px;
-	max-width: 22px;
-	height: 22px;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	padding-right: 2px;
-}
-
-.icon svg {
-	width: 16px;
-	height: 16px;
-}
-
-.label {
-	line-height: 22px;
-	font-size: 13px;
-	cursor: pointer;
-	flex-grow: 1;
-	color: var(--vscode-foreground);
-	text-overflow: ellipsis;
-	overflow: hidden;
-	white-space: nowrap;
-	user-select: none;
-}
-
-.subLabel {
-	line-height: 20px;
-	font-size: 11px;
-	cursor: pointer;
-	flex-grow: 1;
-	color: var(--vscode-foreground);
-	text-overflow: ellipsis;
-	overflow: hidden;
-	white-space: nowrap;
-	user-select: none;
-	text-align: right;
-}
-
-.actions {
-	display: flex;
-	margin: 0 auto;
-	padding: 0;
-	height: 100%;
-	align-items: center;
-	min-width: fit-content;
-}
-
-.actions a {
-	padding: 2px;
-	display: flex;
-	font-size: 11px;
-	padding: 3px;
-	border-radius: 5px;
-	text-decoration: none;
-	cursor: pointer;
-	color: var(--vscode-list-inactiveSelectionForeground);
-}
-
-.actions a:hover {
-	background-color: var(--vscode-toolbar-hoverBackground);
-}
-
-.codicon {
-	flex-shrink: 0;
-	width: 16px;
-	display: flex !important;
-	align-items: center;
-	justify-content: center;
-	font-size: 16px;
-	cursor: pointer;
-}
